+++ src/services/UserProfileAPI.js
@@ -1,14 +1,16 @@
 import request from 'axios';
+import axios from 'axios';
 import Environment from './Environment'
 class UserProfileAPI {
 
-    static usersEndPoint = `${Environment.apiUrl}/user-profiles`;
+    static userProfileEndPoint = `${Environment.apiUrl}/user-profiles`;
+    static userEndPoint = `${Environment.apiUrl}/users`;
 
     static getUsername = (username) => {
 
         const apiCompletionPromise = request({
             method: 'get',
-            url: `${this.usersEndPoint}/user-name/?username=${username}/`,
+            url: `${this.userProfileEndPoint}/user-name/?username=${username}/`,
         });
 
         return apiCompletionPromise;
@@ -18,10 +20,41 @@ class UserProfileAPI {
 
         const apiCompletionPromise = request({
             method: 'get',
-            url: `${this.usersEndPoint}/${userId}/${contentType}/`,
+            url: `${this.userProfileEndPoint}/${userId}/${contentType}/`,
         });
 
         return apiCompletionPromise;
+    }
+
+    static login = (loginCredentials) => {
+
+        const apiCompletionPromise = request({
+            method: 'post',
+            data: loginCredentials,
+            url: `${Environment.apiUrl}/login/`,
+        });
+
+        return apiCompletionPromise;
+    };
+
+    static createUser = (registrationData) => {
+
+        const apiCompletionPromise = request({
+            method: 'post',
+            data: registrationData,
+            url: `${this.userEndPoint}/`,
+        });
+
+        return apiCompletionPromise;
+    };
+
+    static authenticateRequests = (jwtToken, userId) => {
+
+        localStorage.setItem('token', jwtToken);
+        localStorage.setItem('userId', userId);
+        if (jwtToken) {
+            axios.defaults.headers.common['Authorization'] = `JWT ${jwtToken}`;
+        }
     }
 }